Replace deprecated attrs "cmp" argument
The "cmp" argument was deprecated in attrs 19.2.0 in favour of "eq" (see
the changelog at [1]), so switch to using "eq" and update the minimum
required attrs version.

We also need to update the attrs version that MyPy depends upon in the
pre-commit config, so while we are at it update the other pre-commit
package versions as well.

[1] https://www.attrs.org/en/stable/changelog.html