Align step keywords to the longest keyword in the current scenario
When --alignment left or --alignment right is specified, align step
keywords to the longest keyword in the current scenario, instead of the
longest keyword of all keywords for the current Gherkin dialect. This
fixes step keyword alignment for English scenarios with no "Given" step, and
also fixes a problem where the Gherkin dialect was always treated as
English for the purpose of finding step keyword length.

Fixes: #27