Align step text to the longest keyword in the document
When --alignment left or --alignment right is specified, align step text
to the longest keyword in the document, instead of the longest keyword
of all keywords for the current Gherkin dialect. This prevents extra
padding spaces from being displayed for documents that don't use the
longest available keyword. It also fixes a problem where non-English
documents were treated as having English keywords for the purposes of
step keyword alignment.

Fixes: #27