Allow formatting tags on a single line
Add a --single-line-tags option, which causes consecutive tags to be
output on a single line. Also add a --multi-line-tags option to
explicitly specify the current default behaviour of outputting one tag
per line.

When the --single-line-tags option is specified, tags separated by
comments are merged into one line, as discussed in issue #22.