Prevent Gherkin from using feature contents as paths
Subclass the gherkin-official TokenScanner class so that it doesn't
treat the feature contents as a path on the file system. This is
necessary to prevent "path too long for Windows" errors for feature
files larger than 32KB.

Also pin the gherkin-official version to 4.1.3, as we now depend on
Gherkin internals.

Fixes #34